Fix password meter
